atbash-cipher: Simpler atbash encoding and spacing
As per the mentor comments, simplify atbash encoding with a pure
mathematical function.

Also use the nice idea to option<T>.into_iter() to avoid allocating on
the stack intermediate strings or vectors, and also avoid the problem of
having different types of iterators on the arms of the if (when
beginning a new 5 digit block). With plain iterators the if arm would be
std::iter::once, while the else woudl be std::iter::empty.

This version of the code is my 7th version uploaded on exercism.io.

Signed-off-by: Eddy Petrișor <eddy.petrisor@gmail.com>